+++ pub/cmd/bin/scripts.dart
@@ -1,8 +1,14 @@
+part of 'install.dart';
+
 /// Integrating Scripts into the code
 List<String> scripts = [
   'doctor.sh',
   'run.sh',
+  'clean.sh'
   'build.sh',
   'pub.sh',
+  /// Firebase scripts
+  'deploy.sh',
+  'cors_gcs.sh',
   /// Add further scripts here...
 ];